+++ apps/frontend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # frontend
 
+## 1.0.38
+
+### Patch Changes
+
+- 4b901463: SOV-3560: disable permit functionality for DLLR
+- 1d44cce6: SOV-3457: Show treasury proposal executable details
+
+## 1.0.37
+
+### Patch Changes
+
+- cb660fb1: SOV-3390: Fixed Rate Borrow - add ability to increase debt without adding collateral
+- Updated dependencies [e24befac]
+  - @sovryn/ui@1.0.22
+
 ## 1.0.36
 
 ### Patch Changes